Replace optocoupler ballast resistor with 200 Ohm
Replaced the 330 Ohm resistor (used as ballast resistor for the
optocoupler LED) with a 200 Ohm resistor. The initial resistance
calculation was based on a 5V input, overlooking the fact that the D1
Mini pins output at 3.3V.

Special thanks to Zoltán Komló for discovering and troubleshooting the
issue and for proposing the solution.